feat(llc): created a useAutoDimissPostOnboardingEntryPoint hook
