+++ libs/ledger-live-common/src/postOnboarding/hooks/usePostOnboardingEntryPointVisibleOnWallet.test.ts
@@ -5,16 +5,13 @@ import { renderHook } from "@testing-library/react";
 import {
   postOnboardingDeviceModelIdSelector,
   walletPostOnboardingEntryPointDismissedSelector,
-  entryPointFirstDisplayedDateSelector,
 } from "../reducer";
 import { useAllPostOnboardingActionsCompleted } from "./useAllPostOnboardingActionsCompleted";
 import { usePostOnboardingEntryPointVisibleOnWallet } from "./usePostOnboardingEntryPointVisibleOnWallet";
 import { DeviceModelId } from "@ledgerhq/types-devices";
 
-const mockedDispatch = jest.fn();
 jest.mock("react-redux", () => ({
   useSelector: fun => fun(),
-  useDispatch: () => mockedDispatch,
 }));
 jest.mock("../reducer");
 jest.mock("./useAllPostOnboardingActionsCompleted");
@@ -22,25 +19,12 @@ jest.mock("./useAllPostOnboardingActionsCompleted");
 const mockedUseAllCompleted = jest.mocked(useAllPostOnboardingActionsCompleted);
 const mockedDismissedSelector = jest.mocked(walletPostOnboardingEntryPointDismissedSelector);
 const mockedDeviceModelIdSelector = jest.mocked(postOnboardingDeviceModelIdSelector);
-const mockedEntryPointFirstDisplayedDateSelector = jest.mocked(
-  entryPointFirstDisplayedDateSelector,
-);
 
 describe("usePostOnboardingEntryPointVisibleOnWallet", () => {
-  beforeAll(() => {
-    jest.useFakeTimers().setSystemTime(new Date("2020-01-20"));
-  });
-
-  afterAll(() => {
-    jest.runOnlyPendingTimers();
-    jest.useRealTimers();
-  });
-
   beforeEach(() => {
     mockedUseAllCompleted.mockClear();
     mockedDismissedSelector.mockClear();
     mockedDeviceModelIdSelector.mockClear();
-    mockedDispatch.mockClear();
   });
 
   it("should be false if deviceModelId is null", () => {
@@ -84,30 +68,5 @@ describe("usePostOnboardingEntryPointVisibleOnWallet", () => {
     const { result } = renderHook(() => usePostOnboardingEntryPointVisibleOnWallet());
 
     expect(result.current).toBe(true);
-  });
-
-  it("should dismiss the entry point if it has been displayed more than 7 days ago", () => {
-    mockedDismissedSelector.mockReturnValue(false);
-    mockedUseAllCompleted.mockReturnValue(false);
-    mockedDeviceModelIdSelector.mockReturnValue(DeviceModelId.nanoX);
-    mockedEntryPointFirstDisplayedDateSelector.mockReturnValue(new Date("2020-01-12"));
-
-    renderHook(() => usePostOnboardingEntryPointVisibleOnWallet());
-
-    expect(mockedDispatch).toHaveBeenCalledWith({
-      type: "POST_ONBOARDING_HIDE_WALLET_ENTRY_POINT",
-    });
-  });
-
-  it("should be true if the entry point has been displayed less than 7 days ago", () => {
-    mockedDismissedSelector.mockReturnValue(false);
-    mockedUseAllCompleted.mockReturnValue(false);
-    mockedDeviceModelIdSelector.mockReturnValue(DeviceModelId.nanoX);
-    mockedEntryPointFirstDisplayedDateSelector.mockReturnValue(new Date("2020-01-14"));
-
-    const { result } = renderHook(() => usePostOnboardingEntryPointVisibleOnWallet());
-
-    expect(result.current).toBe(true);
-    expect(mockedDispatch).not.toHaveBeenCalled();
   });
 });
